refactor(engines/vllm): factor vLLM weight-update logic into a tier-2 layer
Splits the vLLM refit path into explicit tiers so the WeightTransferEngine
backend is a thin adapter and the generic receiver stays vLLM-free:

- NEW weight_update.py (tier 2): MxVllmWeightUpdater owns the vLLM-specific
  logic — geometry discovery, receive/translate (Megatron + DTensor), buffer
  alloc/registration (host/device/arena), EP filter, byte-verify — behind the
  base lifecycle initialize_weight_update_setup / was_weight_update_setup_
  initialized / start_weight_update / update_weights(subset) / finish_weight_
  update. Info dataclasses (MxInitInfo/MxUpdateInfo, subclassing vLLM's ABC
  bases) + WeightSubset live here. Adds param-name subset scoping.
- weight_transfer.py (tier 3): MxWeightTransferEngine reduces to an adapter
  mapping the ABC's init/receive/shutdown onto the tier-2 lifecycle; re-exports
  the dataclasses for back-compat.
- MxV2RefitReceiver (tier 1) unchanged — stays a generic receiver.

Matches the layering agreed with the NIXL team; a framework can drive tier 2
directly via RPC instead of the ABC. Validated on a live vLLM 0.20.1 pod:
'mx' auto-registers, lifecycle present, shim resolves to the same class.
